Add func to remove duplicate sales on the same day
Some tax lots, for whatever reason, contain multiple sales listed for
the same day, which causes a chain of issues that led to duplicate
buildings listed in the final data: counter_id() used the SaleDate to pick the
highest price sale, which resulted with some buildings having multiple
counter_id = 1, and counter_id was then assumed to be unique.

This function filters the data to have only one sale (the highest
price or the first one if multiple sales for the same day have the same
SalePrice) for each SaleDate.